WIP: Integration of PID

+++ Firmware/aire_apollo/src/MechVentilation.h
@@ -101,7 +101,7 @@ private:
     int _cfgPorcentajeInspiratorio;
 
     /** Pression trigger value in litres per minute. */
-    float _cfgCmh2oTriggerValue;
+    double _cfgCmh2oTriggerValue;
     /**  Ciclo time in seconds. */
     float _cfgSecCiclo;
     /**  Insufflation time in seconds. */
@@ -109,13 +109,13 @@ private:
     /* Exsufflation time in seconds. */
     float _cfgSecTimeExsufflation;
     /* Presion peep (presión mínima en pulmones a la salida). CMH2O */
-    float _cfgPresionPeep;
+    double _cfgPresionPeep;
     /* Presión pico de insuflación CMH2O */
-    float _cfgPresionPico;
+    double _cfgPresionPico;
     /* Presión meseta de insuflación CMH2O */
-    float _cfgPresionMeseta;
+    double _cfgPresionMeseta;
     /* Presión Máxima de seguridad de insuflación CMH2O */
-    float _cfgPresionMax;
+    double _cfgPresionMax;
 
     /** Current state. */
     State _currentState;
